refactor: Create dependencies file. Remove extra files and code parts like propeties and etc. Other fixes according to pr
