+++ auth_app/services/users.py
@@ -1,18 +1,14 @@
 from uuid import UUID
 
 from aiobotocore.client import AioBaseClient
-from fastapi import Depends
 from redis.asyncio.client import Redis
-from sqlalchemy.ext.asyncio import AsyncSession
 
 from auth_app.config import jwt_settings
-from auth_app.db.connect_redis import get_redis_client
 from auth_app.exeptions.custom import (
     ServiceError,
     UserVerificationError,
 )
 from auth_app.messages.common import msg_creator
-from auth_app.middleware.db_session import get_db_from_request
 from auth_app.models import UserORM
 from auth_app.repositories.tokens import TokenRepo
 from auth_app.repositories.users import UserRepo
@@ -24,7 +20,6 @@ from auth_app.schemes.users import (
     PatchUserScheme,
     RoleEnum,
 )
-from auth_app.services.ses.clients import get_ses_client
 from auth_app.services.ses.ses_handler import ses_handler
 from auth_app.services.utils.pwd_hashing import hash_password
 from auth_app.services.utils.verification import verify_auth_code
@@ -47,18 +42,6 @@ class UserService:
     def user_repo(self) -> UserRepo:
         return self.__user_repo
 
-    @property
-    def token_repo(self) -> TokenRepo:
-        return self.__token_repo
-
-    @property
-    def redis(self) -> Redis:
-        return self.__redis
-
-    @property
-    def ses(self) -> AioBaseClient:
-        return self.__ses
-
     async def create_user_record(
         self,
         user_data: CreateUserExtendedScheme,
@@ -69,7 +52,7 @@ class UserService:
                 != jwt_settings.ADMIN_SECRET.get_secret_value()
             ):
                 raise ServiceError("Invalid role or permission code")
-        record = await self.user_repo.create_user(user_data)
+        record = await self.__user_repo.create_user(user_data)
         return record
 
     async def create_init_code_message(
@@ -79,8 +62,8 @@ class UserService:
         email_to = record.email
         await ses_handler.send_confirmation_email(
             email_to=email_to,
-            ses=self.ses,
-            redis_client=self.redis,
+            ses=self.__ses,
+            redis_client=self.__redis,
         )
         response = {
             "record": GetUserScheme.model_validate(record),
@@ -95,8 +78,8 @@ class UserService:
         email_to = payload["email"]
         await ses_handler.send_confirmation_email(
             email_to=email_to,
-            ses=self.ses,
-            redis_client=self.redis,
+            ses=self.__ses,
+            redis_client=self.__redis,
         )
         response = {
             "message": msg_creator.get_code_message(email_to),
@@ -120,7 +103,7 @@ class UserService:
             exclude_unset=True,
             exclude_defaults=True,
         )
-        result = await self.user_repo.update_user(
+        result = await self.__user_repo.update_user(
             user_id=UUID(payload["user_id"]),
             patch_dict=patch_dict,
         )
@@ -135,7 +118,7 @@ class UserService:
         email_to = payload["email"]
         data = await ses_handler.reset_password(
             email_to=email_to,
-            ses=self.ses,
+            ses=self.__ses,
         )
         new_pwd_hash = hash_password(data["new_password"])
         patch_model = PatchUserScheme(password_hash=new_pwd_hash)
@@ -143,7 +126,7 @@ class UserService:
             exclude_unset=True,
             exclude_defaults=True,
         )
-        record = await self.user_repo.update_user(
+        record = await self.__user_repo.update_user(
             user_id=UUID(payload["user_id"]),
             patch_dict=patch_dict,
         )
@@ -153,13 +136,3 @@ class UserService:
             "message": data.get("message"),
         }
         return response
-
-
-async def get_user_service(
-    session: AsyncSession = Depends(get_db_from_request),
-    redis: Redis = Depends(get_redis_client),
-    ses: AioBaseClient = Depends(get_ses_client),
-) -> UserService:
-    user_repo = UserRepo(session)
-    token_repo = TokenRepo(session)
-    return UserService(user_repo, token_repo, redis, ses)
